requirements.txt
django>=1.6,<1.7
django-debug-toolbar==1
django-floppyforms
django-haystack
django-jinja==0.24
django-mptt<0.6
jinja2
markdown<2.3
path.py
Pillow
pytz==2013b
recaptcha-client
South<0.9
Unidecode
whoosh<2.6
coverage
django-nose
django-webtest
factory_boy
fake-factory
